pyproject.toml
[project]
name = "langchain-summarization"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Muhammad Mehdi", email = "Mehdi.Abidi@imperiumdynamics.com"},
]
dependencies = ["python-dotenv==1.2.2", "langchain==0.1.20", "langchain-openai==0.1.7", "openai==1.109.1", "tiktoken==0.12.0", "langchain-core==0.1.53", "langchain-community==0.0.38", "langchain-text-splitters==0.0.2", "faiss-cpu==1.13.2", "numpy==1.26.4", "pypdf==6.9.2", "beautifulsoup4==4.14.3", "requests==2.33.0", "PyYAML==6.0.3", "tqdm==4.67.3"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
